got rid of unnecessary 'watch' method in hook form
