+++ Frontend/src/Components/registerPage/interfaces/BtnProps.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react'
+
+interface btnProps {
+  children: ReactNode
+}
+
+export default btnProps
