+++ Frontend/src/Components/registerPage/interfaces/IFormInputs.ts
@@ -1,0 +1,9 @@
+interface IFormInputs {
+  name: string
+  email: string
+  password: string
+  passwordConfirm: string
+  acceptTerms: boolean
+}
+
+export default IFormInputs
